refactor: Add fix to nav bar search
