+++ src/Components/ArticleDetail/ArticleDetail.css
@@ -11,7 +11,7 @@
     margin: 10px 0;
     padding: 5px 15px;
     background-color: #333;
-    color: #fff;
+    color: #de0f0f;
     border: none;
     cursor: pointer;
     transition: background-color 0.2s;
