+++ src/Components/ArticleList/ArticleList.css
@@ -1,4 +1,4 @@
-body {
+/* body {
     font-family: 'Georgia', serif;
     background-color: #f3f3f3;
     color: #1c1c1c;
@@ -21,6 +21,8 @@ body {
     cursor: pointer;
     border-radius: 3px;
     font-size: 0.9rem;
+    order: 1;
+    margin-right: 20px;
 }
 
 .switch-button:hover {
@@ -43,7 +45,6 @@ body {
     justify-content: center;
     gap: 20px;
     width: 100%;
-    margin-left: 20rem;
 }
 
 .article-link {
@@ -84,55 +85,6 @@ body {
     word-wrap: break-word;
     margin: 5px 0;
     padding: 0 5px;
-}
-
-.search-container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 5px;
-    background-color: #444;
-    margin-bottom: 20px;
-    border-radius: 5px;
-    box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.1);
-    margin-top: -50px;
-}
-
-.search-input, #categorySelect {
-    border: none;
-    border-radius: 20px;
-    background-color: rgba(255, 255, 255, 0.1);
-    color: #fff;
-    transition: background-color 0.3s;
-    outline: none;
-    font-weight: bold;
-    flex: 1;
-    padding: 5px 10px;
-    font-size: 14px;
-    margin: 0 3px;
-}
-
-.search-button {
-    background-color: #e50914;
-    color: #fff;
-    border: none;
-    border-radius: 0 20px 20px 0;
-    padding: 5px 10px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-    font-weight: bold;
-    outline: none;
-    display: flex;
-    align-items: center;
-}
-
-.search-button:hover {
-    background-color: #f40612;
-}
-
-.search-button::before {
-    content: "\1F50D";  
-    margin-right: 5px;
 }
 
 .article-detail-container {
@@ -178,6 +130,35 @@ body {
     margin: 0 auto;
 }
 
+.clear-button {
+    background: none;
+    border: none;
+    font-size: 20px;
+    cursor: pointer;
+    color: #aaa;
+    position: absolute;
+    right: 10px;
+    top: 50%;
+    transform: translateY(-50%);
+}
+
+@media (max-width: 600px) {
+    .controls-container {
+        flex-direction: column;
+        align-items: stretch;
+    }
+
+    .search-container {
+        order: 1;
+        margin-bottom: 10px;
+    }
+
+    .switch-button {
+        order: 2;
+        margin-right: 0;
+    }
+}
+
 @media (max-width: 1000px) {
     .article-link {
         flex: 1 1 calc(50% - 20px);
@@ -185,75 +166,76 @@ body {
     }
 }
 
-@media (max-width: 600px) {
-    .article-link {
-        flex: 1 1 100%;
-        max-width: 100%;
-    }
-
-    .article-list-container {
-        margin-left: 0;
-    }
+ */
+ 
+ .switch-button {
+    border: none;
+    padding: 0.5rem 1rem;
+    margin: 1rem 0;
+    cursor: pointer;
+    border-radius: 3px;
+    font-size: 0.9rem;
+    order: 1;
+    margin-right: 20px;
 }
 
-/* 
-.navbar-search {
+.switch-button:hover {
+    opacity: 0.9;
+}
+
+.switch-button-live {
+    background-color: #2196F3;
+    color: #ffffff;
+}
+
+.switch-button-mock {
+    background-color: #0ecb3a;
+    color: #ffffff;
+}
+
+.article-list-container {
+    padding: 20px;
     display: flex;
-    align-items: center;
-}
-
-.navbar-search input {
-    padding: 5px 10px;
-    border: 1px solid #ddd;
-    outline: none;
-    margin-right: 10px;
-    transition: border 0.3s;
-    border-radius: 0 4px 4px 0;
-}
-
-.navbar-search input:focus {
-    border-color: #e50914;
-}
-
-.navbar-search button {
-    padding: 5px 15px;
-    background-color: #e50914;
-    color: #fff;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.2s;
-}
-
-.navbar-search button:hover {
-    background-color: #f40612;
-}
-
-.navbar-search select {
-    padding: 5px 10px;
-    border: 1px solid #ddd;
-    border-radius: 4px 0 0 0;
-    margin-right: 10px;
-    outline: none;
-    transition: border 0.3s;
-    cursor: pointer;
-}
-
-.navbar-search select:focus {
-    border-color: #e50914;
-}
-
-.navbar-search button[type="button"] {
-    background-color: transparent;
-    color: #e50914;
-    padding-left: 5px;
-    padding-right: 5px;
-    border: none;
-}
-
-@media (max-width: 768px) {
-    .navbar-search {
-        width: 100%;
-        justify-content: center;
-    }
-} */
+    flex-wrap: wrap;
+    gap: 20px;
+    justify-content: center;
+  }
+  
+  .article-link {
+    display: flex;
+    flex-direction: column;
+    border: 1px solid #e5e5e5;
+    width: 300px;
+    border-radius: 10px;
+    overflow: hidden;
+    text-decoration: none;
+    color: inherit;
+    transition: transform 0.3s;
+  }
+  
+  .article-link:hover {
+    transform: scale(1.05);
+  }
+  
+  .article-link img {
+    width: 100%;
+    height: 200px;
+    object-fit: cover;
+    border-bottom: 1px solid #e5e5e5;
+  }
+  
+  .article-text-content {
+    padding: 10px;
+  }
+  
+  .article-text-content h2 {
+    font-size: 18px;
+    margin-bottom: 10px;
+  }
+  
+  .article-text-content p {
+    font-size: 14px;
+    margin-bottom: 5px;
+  }
+  
+
