+++ src/Components/ArticleList/ArticleList.tsx
@@ -15,6 +15,7 @@
 //   };
 //   publishedAt: string;
 //   content: string;
+//   category: string;
 // };
 
 // interface Props {
@@ -28,6 +29,8 @@
 //   const [loading, setLoading] = useState(true);
 //   const [error, setError] = useState<string | null>(null);
 //   const [retryCount, setRetryCount] = useState(0);
+//   const [selectedCategory, setSelectedCategory] = useState('All');
+//   const categories = ['All', 'Business', 'Entertainment', 'Health', 'Science', 'Sports', 'Technology'];
 
 //   useEffect(() => {
 //     fetchNews("bitcoin", 1, useMockData)
@@ -41,60 +44,76 @@
 //         setLoading(false);
 //       });
 //   }, [useMockData, retryCount]);
-
+  
 //   const filteredArticles = articles.filter(article => 
-//     article.title.toLowerCase().includes(searchTerm.toLowerCase())
+//     (selectedCategory === 'All' || article.category === selectedCategory) && 
+//     (!searchTerm || article.title.toLowerCase().includes(searchTerm.toLowerCase()))
 //   );
 
-//   if (loading) return <div>Loading...</div>;
-//   if (error) return (
-//     <div>
-//       {error}
-//       <button onClick={() => setRetryCount(retryCount + 1)}>Retry</button>
-//     </div>
-//   );
+//   const handleSearchChange = (e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
+//     setSearchTerm(e.target.value);
+//   };
 
 //   return (
 //     <div className="article-list-container">
 //       <div className="search-container">
-//         <label htmlFor="searchInput" className="visually-hidden">Search Articles</label>
+//         <select
+//           id="categorySelect"
+//           name="categorySelect"
+//           onChange={handleSearchChange}
+//           value={searchTerm}
+//         >
+//           <option value="">Type or select a category...</option>
+//           {categories.map(category => (
+//             <option key={category} value={category}>
+//               {category}
+//             </option>
+//           ))}
+//         </select>
 //         <input
 //           id="searchInput"
 //           name="searchInput"
 //           type="text"
 //           className="search-input"
-//           placeholder="Search articles..."
+//           placeholder="Type or select a category..."
 //           value={searchTerm}
-//           onChange={(e) => setSearchTerm(e.target.value)}
+//           onChange={handleSearchChange}
 //         />
+//         <button 
+//           className={`clear-button ${searchTerm ? 'visible' : 'hidden'}`} 
+//           onClick={() => setSearchTerm('')}
+//         >
+//           X
+//         </button>
 //       </div>
-//       <div className="articles">
-//         {filteredArticles.map((article, index) => (
-//           <Link to={`/article/${index}`} key={index} className="article-link">
+//       {filteredArticles.map((article, index) => (
+//         <Link to={`/article/${index}`} key={index} className="article-link">
 //           <img 
-//               src={article.urlToImage || missingImg} 
-//               onError={(e) => {
-//                   const imgElement = e.target as HTMLImageElement;
-//                   imgElement.onerror = null;
-//                   imgElement.src = missingImg;
-//               }} 
-//               alt={article.title || "No Title"} 
+//             src={article.urlToImage || missingImg} 
+//             onError={(e) => {
+//               const imgElement = e.target as HTMLImageElement;
+//               imgElement.onerror = null;
+//               imgElement.src = missingImg;
+//             }} 
+//             alt={article.title || "No Title"} 
 //           />
 //           <div className="article-text-content">
-//               <h2>{article.title}</h2>
-//               <p>{article.description}</p>
-//               <p>{new Date(article.publishedAt).toLocaleDateString()}</p>
-//               <p>{article.source.name}</p>
+//             <h2>{article.title}</h2>
+//             <p>{article.description}</p>
+//             <p>{new Date(article.publishedAt).toLocaleDateString()}</p>
+//             <p>{article.source.name}</p>
 //           </div>
-//       </Link>
-      
-//         ))}
-//       </div>
+//         </Link>
+//       ))}
 //     </div>
 //   );
 // }
 
 // export default React.memo(ArticleList);
+
+
+// ArticleList.tsx
+
 import React, { useState, useEffect } from 'react';
 import { fetchNews } from '../../apiCalls';
 import '../ArticleList/ArticleList.css';
@@ -112,6 +131,7 @@ export type Article = {
   };
   publishedAt: string;
   content: string;
+  category: string;
 };
 
 interface Props {
@@ -121,13 +141,9 @@ interface Props {
 
 const ArticleList: React.FC<Props> = ({ useMockData }) => {
   const [articles, setArticles] = useState<Article[]>([]);
-  const [searchTerm, setSearchTerm] = useState('');
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
-  const [tempSearchTerm, setTempSearchTerm] = useState('');
   const [retryCount, setRetryCount] = useState(0);
-  const [selectedCategory, setSelectedCategory] = useState('All');
-  const categories = ['All', 'Business', 'Entertainment', 'Health', 'Science', 'Sports', 'Technology'];
 
   useEffect(() => {
     fetchNews("bitcoin", 1, useMockData)
@@ -142,61 +158,27 @@ const ArticleList: React.FC<Props> = ({ useMockData }) => {
       });
   }, [useMockData, retryCount]);
 
-  const filteredArticles = articles.filter(article => 
-    article.title.toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
-  const handleSearch = () => {
-    setSearchTerm(tempSearchTerm);
-  };
-
   return (
     <div className="article-list-container">
-      <div className="search-container">
-        <select
-          id="categorySelect"
-          name="categorySelect"
-          onChange={(e) => setSelectedCategory(e.target.value)}
-        >
-          <option value="">Search or select a category...</option>
-          {categories.map(category => (
-            <option key={category} value={category}>
-              {category}
-            </option>
-          ))}
-        </select>
-        <input
-          id="searchInput"
-          name="searchInput"
-          type="text"
-          className="search-input"
-          placeholder="Search articles..."
-          value={tempSearchTerm} // Use tempSearchTerm here
-          onChange={(e) => setTempSearchTerm(e.target.value)} // Update tempSearchTerm as user types
-        />
-        <button onClick={handleSearch} className="search-button">Search</button> {/* New search button */}
-      </div>
-      <div className="articles">
-        {filteredArticles.map((article, index) => (
-          <Link to={`/article/${index}`} key={index} className="article-link">
-            <img 
-              src={article.urlToImage || missingImg} 
-              onError={(e) => {
-                const imgElement = e.target as HTMLImageElement;
-                imgElement.onerror = null;
-                imgElement.src = missingImg;
-              }} 
-              alt={article.title || "No Title"} 
-            />
-            <div className="article-text-content">
-              <h2>{article.title}</h2>
-              <p>{article.description}</p>
-              <p>{new Date(article.publishedAt).toLocaleDateString()}</p>
-              <p>{article.source.name}</p>
-            </div>
-          </Link>
-        ))}
-      </div>
+      {articles.map((article, index) => (
+        <Link to={`/article/${index}`} key={index} className="article-link">
+          <img 
+            src={article.urlToImage || missingImg} 
+            onError={(e) => {
+              const imgElement = e.target as HTMLImageElement;
+              imgElement.onerror = null;
+              imgElement.src = missingImg;
+            }} 
+            alt={article.title || "No Title"} 
+          />
+          <div className="article-text-content">
+            <h2>{article.title}</h2>
+            <p>{article.description}</p>
+            <p>{new Date(article.publishedAt).toLocaleDateString()}</p>
+            <p>{article.source.name}</p>
+          </div>
+        </Link>
+      ))}
     </div>
   );
 }
