+++ src/Components/NavBar/NavBar.css
@@ -1,3 +1,101 @@
+/* .navbar {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 10px 20px;
+  background-color: #333;
+  color: #fff;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  transition: all 0.3s;
+}
+
+.navbar-logo {
+  height: 50px;
+  width: auto;
+  cursor: pointer;
+}
+
+.navbar-name {
+  font-size: 24px;
+  margin-left: 20px;
+  flex: 1;
+}
+
+.navbar-search {
+  display: flex;
+  align-items: center;
+}
+
+.navbar-search input {
+  padding: 5px 10px;
+  border: 1px solid #ddd;
+  outline: none;
+  margin-right: 10px;
+  transition: border 0.3s;
+  border-radius: 0 4px 4px 0;
+}
+
+.navbar-search input:focus {
+  border-color: #e50914;
+}
+
+.navbar-search button {
+  padding: 5px 15px;
+  background-color: #e50914;
+  color: #fff;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+}
+
+.navbar-search button:hover {
+  background-color: #f40612;
+}
+
+.navbar-search select {
+  padding: 5px 10px;
+  border: 1px solid #ddd;
+  border-radius: 4px 0 0 0;
+  margin-right: 10px;
+  outline: none;
+  transition: border 0.3s;
+  cursor: pointer;
+}
+
+.navbar-search select:focus {
+  border-color: #e50914;
+}
+
+.navbar-search button[type="button"] {
+  background-color: transparent;
+  color: #e50914;
+  padding-left: 5px;
+  padding-right: 5px;
+  border: none;
+}
+
+@media (max-width: 768px) {
+  .navbar {
+      flex-direction: column;
+      align-items: stretch;
+  }
+
+  .navbar-name {
+      margin: 10px 0;
+      text-align: center;
+  }
+
+  .navbar-search {
+      width: 100%;
+      justify-content: center;
+  }
+} */
+
+
+
+/* NavBar.css */
+
 .navbar {
   display: flex;
   align-items: center;
